[WIP] Migrating Page component over to typescript
