+++ src/components/Page/index.ts
@@ -1,2 +1,1 @@
-// @flow strict
 export { default } from './Page';
